Renamed reliability.h to rmodel.h
